// api/playauto_api_type.ts
export const shopDataNameInfo: { [key: string]: string } = {
    "A001": "옥션",
    "A522": "옥션2.0",
    "A006": "G마켓",
    "A523": "G마켓2.0",
    "A077": "스마트스토어",
    "A027": "인터파크 - 오픈마켓",
    "A112": "11번가",
    "A113": "11번가(단일상품)",
    "B955": "마이소",
    "B118": "메이크샵",
    "B378": "쿠팡",
    "B956": "티몬",
    "B430": "위메프",
    "B719": "위메프2.0",
    "C003": "Qoo10(sg)",
    "C010": "Qoo10(jp)",
    "A088": "멸치쇼핑",
    "B579": "NH마켓",
    "A011": "GS SHOP",
    "A012": "롯데닷컴",
    "A004": "CJmall",
    "B008": "half club",
    "A131": "현대H몰 (Hmall)",
    "B009": "LOTTE DEPARTMENT STORE",
    "B617": "오가게",
    "A017": "Akmall",
    "A065": "iSTYLE 24",
    "A024": "WIZWID",
    "B394": "the galleria",
    "A118": "ns홈쇼핑",
    "B666": "MUSINSA",
    "B733": "PLAYER",
    "B612": "LOTTE MART",
    "B614": "홈앤쇼핑",
    "B007": "패션플러스",
    "B957": "W컨셉",
    "A032": "SSG(통합) - 신세계몰, SSG(이마트몰), 이마트트레이더스몰",
    "B609": "다이소몰",
    "B048": "1300k",
    "B041": "텐바이텐",
    "B051": "바보사랑",
    "B599": "인터파크 도서",
    "B046": "예스24",
    "B888": "후추통",
    "B572": "골핑(B2C)",
    "B087": "sbs골프닷컴",
    "A132": "현대H몰(백화점)",
    "B325": "이지웰",
    "A085": "SSG(신세계백화점)",
    "B701": "삼성카드쇼핑",
    "B690": "K쇼핑",
    "B771": "제로투세븐",
    "A100": "롯데아이몰",
    "B661": "이랜드몰",
    "B693": "도매꾹",
    "B668": "도매창고",
    "B730": "오너클랜",
    "B669": "머스트잇",
    "B682": "브리치",
    "B685": "공영홈쇼핑",
    "B688": "카카오톡스토어",
    "B953": "POOM",
    "B694": "핫트랙스",
    "A133": "더현대닷컴",
    "B696": "하이마트",
    "B502": "홈플러스",
    "B959": "오늘의집",
    "A524": "롯데온",
    "B880": "펀앤쇼핑",
}
// ...
export interface IPAJobCallbackProcessingResponse {
    job_id: string
    current: number
    total: number
    state_msg: string
    job_cate: string
    sol_code: string
}
export interface IPAJobCallbackDoneResponse<T> {
    job_id: string
    title: string
    results: IPAJobCallbackResults<T>
}

export type IPAJobCallbackResponse<T> = IPAJobCallbackProcessingResponse | IPAJobCallbackDoneResponse<T>;

export interface IPAJobCallbackResults<T> {
    "config.json": IPAJobCallbackConfigJson
    "result.json": T[] | IPAJobCallbackFailedResultJson
}

export interface IPAJobCallbackFailedResultJson {
    Result: string
    Message: string
}

export interface IPAJobCallbackConfigJson {
    sol_code: string
}
// ...
export interface IPAJobCallbackScrapOrderResultJson {
    wid: string | null
    site_code: string
    site_id: string
    tid: string
    wdate: string
    odate: string
    cdate: string
    ddate: string
    state_date: string
    deli_date: string
    remit_date: string
    state: number
    state_old: number
    state_msg: any
    ocode: string
    pay_method: number
    pcode: string
    pcode2: string
    pname: string
    pname2: string
    src_pname: string
    popt: string
    src_popt: string
    popt_plus: string
    popt_plus_price: number
    notice_msg: string
    wprice1: number
    wprice2: number
    price: number
    price1: number
    count: number
    deliv_method: number
    deliv_price: number
    oaid: string
    oname: string
    otel: string
    ohtel: string
    oemail: string
    gname: string
    gtel: string
    ghtel: string
    gzip: string
    gaddress: string
    msg: string
    msg2: string
    msg3: string
    fee: any
    tax_doc: any
    sender: string
    sendno: string
    master_code: string
    master_sdate: string
    master_edate: string
    master_stock_check: string
    master_etc1: any
    master_etc2: any
    master_etc3: any
    cancel_flag: number
    misc1: string
    misc2: string
    misc3: string
    misc4: string
    misc5: string
    misc6: string
    misc7: string
    misc8: string
    misc9: string
    misc10: string
    black: number
    outskirts: number
    airpost: number
    white: number
    match_popt: any
    match_popt_plus: any
    match_popt_ok: number
    pcode3: string
    misc11: string
    misc12: string
    misc13: string
    misc14: string
    misc15: string
    misc16: string
    misc17: string
    misc18: string
    misc19: string
    misc20: string
    customer: string
    popt_price: number
    msg_card: string
    gprivate_no: string
    gname_eng: string
    bonus: string
    shopInfo: IPAScrapOrderShopInfo
    ordinfo: any[]
    wprice: number
    add_price1: number
    add_price2: number
    sale_price: number
    tax_price: number
    qxpress: string
    sku_code: string
    invenopt: string
    currency: string
    leaders_status: any
    leaders_price: number
    priceOfB2B: number
    uniq_order_where: IPAScrapOrderUniqOrderWhere
}

export interface IPAScrapOrderShopInfo {
    position: string
    overlap_mode: string
    uniqKey: string
}

export interface IPAScrapOrderUniqOrderWhere { }


export const PlayAutoDelivMethod: { [key: number]: string } = {
    0: "무료배송",
    1: "선결제",
    2: "착불",
    3: "조건부배송",
    4: "방문수령",
    5: "퀵배송",
    6: "일반우편",
    7: "설치배송",
    8: "기타",
}